JIP-2: Misc improvements, additions, and fixes
This primarily fixes a number of poorly defined and out-of-date parts of
the spec. It should now match the RPC interface implemented by PolkaJam
and Parity's JAM tooling (as of 2025-10-31). In particular:

- The operation of subscriptions is properly defined; the JSON-RPC spec
  does not define this. There is not intended to be any functional
  change here.
- The result of bestBlock() and similar methods has been fixed; for a
  long time now bestBlock() has returned an Object with "header_hash"
  and "slot" members rather than an Array.
- A number of cases where methods were defined to return Null if e.g. a
  block is not known have been fixed. An error should be returned
  instead.
- The parameters to the various subscribe methods are now documented.
- The parameters returned by the parameters() method have been updated.
  Note that these parameters should exactly match the parameters
  returned by the PVM fetch host-call, although they are returned as a
  JSON object rather than being encoded using the GP codec.

A number of new methods have been added. These are necessary to e.g. run
the CoreVM builder:

- workReport.
- submitWorkPackageBundle.
- workPackageStatus.
- subscribeWorkPackageStatus.
- fetchWorkPackageSegments.
- fetchSegments.
- syncState.
- subscribeSyncStatus.

Error codes are now defined. These ought to be followed to avoid
confusing clients.

Finally, the formatting of the methods section in particular has
(hopefully) been improved.